+++ lib/eth/src/utils.rs
@@ -16,12 +16,14 @@ pub fn read_abi(path: String) -> Result<Abi, Box<dyn Error>> {
 pub async fn get_block_timestamp<T: JsonRpcClient + Clone>(
     provider: Provider<T>,
     block_hash: H256,
-) -> Result<i64, Box<dyn Error>> {
+) -> Result<u32, Box<dyn Error>> {
     let block = loop {
         match provider.get_block(block_hash).await {
             Ok(Some(block)) => break block,
             Ok(None) => return Err("Block not found".into()),
             Err(e) => {
+                // Retry after 250ms when rate limit hit
+                // This is a temporary solution until we build a custom Provider implementation with native retries
                 if e.to_string().contains("429") {
                     tokio::time::sleep(tokio::time::Duration::from_millis(250)).await;
                     continue;
